+++ common/src/main/java/com/minecraft/job/common/review/domain/Review.java
@@ -61,7 +61,7 @@ public class Review {
         return team.getLeaderEmail();
     }
 
-    public Long getTeamOfAveragePoint() {
+    public Double getTeamOfAveragePoint() {
         return team.getAveragePoint();
     }
